feat: Add derive macro for PrefabData
