+++ amethyst_animation/Cargo.toml
@@ -19,6 +19,7 @@ travis-ci = { repository = "amethyst/amethyst" }
 [dependencies]
 amethyst_assets = { path = "../amethyst_assets/", version = "0.5.0" }
 amethyst_core = { path = "../amethyst_core/", version = "0.4.0" }
+amethyst_derive = { path = "../amethyst_derive", version = "0.1.0" }
 amethyst_renderer = { path = "../amethyst_renderer/", version = "0.9.0" }
 derivative = "1.0"
 fnv = "1"
