+++ amethyst_animation/src/lib.rs
@@ -50,6 +50,8 @@
 
 extern crate amethyst_assets;
 extern crate amethyst_core;
+#[macro_use]
+extern crate amethyst_derive;
 extern crate amethyst_renderer;
 #[macro_use]
 extern crate derivative;
