+++ amethyst_derive/Cargo.toml
@@ -18,7 +18,9 @@ travis-ci = { repository = "amethyst/amethyst" }
 syn = { version = "0.15", features = ["visit"] }
 quote = "0.6"
 proc-macro2 = "0.4"
-amethyst_core = { path = "../amethyst_core", version = "0.4.0" }
+
+[dev-dependencies]
+amethyst_core = { path = "../amethyst_core", version = "0.3.0" }
 
 [lib]
 name = "amethyst_derive"
